requirements.txt
streamlit>=1.35.0
sentence-transformers>=2.7.0
scikit-learn>=1.4.0
pymupdf>=1.24.0
python-docx>=1.1.0
pandas>=2.2.0
